Update for change of int -> long in Output interface.  int -> unsigned gcc compiler warning
git-svn-id: http://svn.code.sf.net/p/protomol/svn/trunk@2106 ce46b6ef-3445-0410-911d-c47028e6f9df

+++ protomol/src/protomol/output/OutputFAHGUI.cpp
@@ -68,7 +68,7 @@ void OutputFAHGUI::doInitialize() {
 }
 
 
-void OutputFAHGUI::doRun(int step) {
+void OutputFAHGUI::doRun(long step) {
   GUIServer::request_t request = server->getRequest();
 
   // GuiPause?
@@ -116,7 +116,7 @@ void OutputFAHGUI::doRun(int step) {
     THROW("GUI communication timeout.");
 }
 
-void OutputFAHGUI::doFinalize(int step) {
+void OutputFAHGUI::doFinalize(long step) {
   doRun(step);
   if (server) {
     delete server;
